test exception

+++ src/main/java/fr/sorbonne/paris/nord/university/api/controller/TeamController.java
@@ -3,6 +3,7 @@ package fr.sorbonne.paris.nord.university.api.controller;
 
 import fr.sorbonne.paris.nord.university.api.dto.TeamDto;
 import fr.sorbonne.paris.nord.university.api.entity.TeamEntity;
+import fr.sorbonne.paris.nord.university.api.exception.TeamInvalidException;
 import fr.sorbonne.paris.nord.university.api.mapper.TeamMapper;
 import fr.sorbonne.paris.nord.university.api.service.TeamService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -31,39 +32,66 @@ public class TeamController {
 
 	@GetMapping("/teams")
 	 public ResponseEntity<List<TeamDto>> getAllTeams() {
+
 		List<TeamDto> teamDto = this.teamService.getAllTeam()
 				.stream()
 				.map(teamMapper::toDto)
 				.collect(toList());
 
-		return new ResponseEntity<>(teamDto, HttpStatus.OK);
+		if (teamDto.isEmpty()) {
+			throw new TeamInvalidException("Liste d'équipe vide");
+		}else{
+			return new ResponseEntity<>(teamDto, HttpStatus.OK);
+		}
+
 	 }
 
 	 @GetMapping("/teams/{id}")
 	public ResponseEntity<TeamDto> getTeamById(@PathVariable Long id){
-		TeamDto teamDto = this.teamService.getTeamById(id)
-				.stream()
-				.map(teamMapper::toDto)
-				.findFirst().get();
 
-		 return new ResponseEntity<>(teamDto, HttpStatus.OK);
+		 if (teamService.getTeamById(id).isPresent()) {
+			 TeamDto teamDto = this.teamService.getTeamById(id)
+					 .stream()
+					 .map(teamMapper::toDto)
+					 .findFirst().get();
+
+			 return new ResponseEntity<>(teamDto, HttpStatus.OK);
+		 } else {
+			 throw new TeamInvalidException("Équipe introuvable avec l'ID : " + id);
+		 }
 	 }
 
 	@PostMapping("/teams/add")
 	 public ResponseEntity<TeamDto> newTeam(@RequestBody TeamDto teamDto){
-		 TeamEntity teamEntity = teamMapper.toTeamEntity(teamDto);
-		 this.teamService.saveTeam(teamEntity);
-		 return new ResponseEntity<TeamDto>(teamDto, HttpStatus.CREATED);
+
+		if (teamDto.getName().isEmpty() || teamDto.getSlogan().isEmpty()) {
+			throw new TeamInvalidException("Le champ 'slogan' ou 'name' ne peut pas être null ou vide.");
+		}else{
+			TeamEntity teamEntity = teamMapper.toTeamEntity(teamDto);
+			this.teamService.saveTeam(teamEntity);
+			return new ResponseEntity<TeamDto>(teamDto, HttpStatus.CREATED);
+		}
+
 	 }
 
 	 @DeleteMapping("/teams/delete/{id}")
 	 public ResponseEntity<String> deleteTeam(@PathVariable Long id){
-		this.teamService.deleteTeam(id);
 
-		String response = "L'équipe avec id "+id+ " a été supprimé";
+		 if (teamService.getTeamById(id).isPresent()) {
 
-		return new ResponseEntity<>(response, HttpStatus.OK);
+			 this.teamService.deleteTeam(id);
+			 String response = "L'équipe avec id "+id+ " a été supprimé";
+
+			 return new ResponseEntity<>(response, HttpStatus.OK);
+		 } else {
+			 throw new TeamInvalidException("Équipe introuvable avec l'ID : " + id);
+		 }
 	 }
+
+	@ExceptionHandler(TeamInvalidException.class)
+	public ResponseEntity<String> handleTeamInvalidException(TeamInvalidException e) {
+		return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(e.getMessage());
+	}
 
 
 }
